pyproject.toml
[project]
name = "taskflow-api"
version = "0.1.0"
description = "TaskFlow API"
authors = [
    {name = "Lifrov Vadim",email = "lvadim3110@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]
